requirements.txt
opencv-python
numpy
pillow
flask
pytesseract
imwatermark
pymupdf
tensorflow==2.15.0
#keras
python-docx
sentence-transformers
nltk
matplotlib
piexif
gunicorn